feat(combatant): relocate automated combat-resume to CombatantLogic + rewire chat dispatch
The defender's automated-combat defenses are a combatant concern; move them off
the actor and dispatch the attack card's defense buttons to the combatant.

- Move automatedBlock/Dodge/Counterstrike/IgnoreResume (+ rehydrateAttackResult/
  buildCombatResult/postCombatResultCard) from BeingLogic to CombatantLogic as
  intrinsic actions. `this` is the defender combatant; `this.actorLogic` supplies
  strike-mode capability. BeingLogic keeps only the attacker entry
  (automatedCombatStart); its dead combat imports are removed.
- SohlCombatant gains `onChatCardButton`, dispatching generically to its logic
  (mirrors SohlActor). The attack card's defense-button handler now addresses the
  defender's **combatant** uuid; chat-card-gating resolves the actor through the
  combatant (statuses/capability via `defender.actor`).
- Combat-result **injury** addresses stay the actor uuid (createInjury runs on
  the actor), via `this.combatant.actor.uuid`.
- SohlLogic.actor resolves the owning actor from the document (handles
  combatant/effect, no longer throws via `this.item`).
- Tests: new CombatantLogic.test covers the resume warn-paths + intrinsic
  actions; the resume describe is removed from Being.test; logicHarness gains
  makeCombatantLogic.

Full suite green (1200 + 59 purity), build/lint/docs clean. NOTE: the
actor->combatant handler flip is runtime-behavioral — verify the defense
buttons in Foundry before merge.
