+++ src/document/actor/logic/BeingLogic.ts
@@ -16,8 +16,6 @@ import type { SuccessTestResult } from "@src/domain/result/SuccessTestResult";
 import type { OpposedTestResult } from "@src/domain/result/OpposedTestResult";
 import type { SohlActionContext } from "@src/core/SohlActionContext";
 import { CombatResult } from "@src/domain/result/CombatResult";
-import { DefendResult } from "@src/domain/result/DefendResult";
-import type { AttackResult } from "@src/domain/result/AttackResult";
 import {
     SohlActorBaseLogic,
     type SohlActorData,
@@ -43,25 +41,7 @@ import {
     selectActorCombatant,
 } from "@src/document/actor/logic/token-helpers";
 import { getActiveScene, getActiveCombat } from "@src/core/FoundryHelpers";
-import {
-    startAutomatedAttackFromActor,
-    showDefenseDialog,
-    findCombatant,
-    resolveCounterstrikeContext,
-    buildAimChoices,
-    pickChoice,
-} from "@src/document/actor/logic/automated-combat";
-import {
-    buildCombatCardData,
-    buildAttackResult,
-    collectAttackableStrikeModes,
-    resolveSkillMasteryLevel,
-    collectBlockableStrikeModes,
-    indexOfBestMastery,
-} from "@src/document/actor/logic/combat-actions";
-import type { MasteryLevelModifier } from "@src/domain/modifier/MasteryLevelModifier";
-import { resolveActionInput } from "@src/utils/actionInput";
-import { instanceFromJSON, toFilePath } from "@src/utils/helpers";
+import { startAutomatedAttackFromActor } from "@src/document/actor/logic/automated-combat";
 import type { SohlTokenDocument } from "@src/document/token/SohlTokenDocument";
 import type { SohlCombatant } from "@src/document/combatant/SohlCombatant";
 import { readBaseMove } from "@src/domain/movement/move-helpers";
@@ -70,11 +50,9 @@ import {
     defineType,
     ITEM_KIND,
     MovementMedium,
-    SKILL_CODE,
     SOHL_ACTION_SCOPE,
     SOHL_CONTEXT_MENU_SORT_GROUP,
     TEST_TYPE,
-    VALUE_DELTA_INFO,
 } from "@src/utils/constants";
 import { SohlAction } from "@src/domain/action/SohlAction";
 import { SimpleRoll } from "@src/utils/SimpleRoll";
@@ -585,478 +563,6 @@ export class BeingLogic<
     }
 
     /**
-     * Present a dialog asking the player to select a strike mode to block with to resume
-     * the automated combat.
-     *
-     * @remarks
-     * Any strike mode that has the noBlock trait should be filtered out of the choices.
-     * The default value of the choices should be the most recently used strike mode that
-     * does not have the noBlock trait. Otherwise, there is no default value.
-     *
-     * Once a strike mode is selected, delegate processing of the block resume to that
-     * strike mode's item.
-     *
-     * One of `combatResult` or `attackResult` must be supplied in `context.scope`:
-     * - `combatResult` is the prior automated resume result that is being reassessed
-     * - `attackResult` is the result of the automated attack that initiated the automated resume
-     *
-     * @param context - Action context carrying the combat or attack result in its scope.
-     * @param [context.scope.priorTestResult] A prior opposed test result that is being retried.
-     * @param [context.scope.attackResult] The test result that initiated the opposed test
-     */
-    async automatedBlockResume(
-        context: SohlActionContext<Partial<CombatResult.ContextScope>>,
-    ): Promise<void> {
-        const attackResult = this.rehydrateAttackResult(context);
-        if (!attackResult) return;
-
-        const entries = collectBlockableStrikeModes(this);
-        if (!entries.length) {
-            sohl.log.uiWarn(
-                `${this.name} has no strike mode able to block.`,
-            );
-            return;
-        }
-        const choices: Record<string, string> = {};
-        entries.forEach((e, i) => {
-            choices[String(i)] = e.label;
-        });
-
-        // Default to the most-recently-used block mode (if still available),
-        // else the best-chance block (highest effective block ML).
-        const combatant = context.token ? findCombatant(context.token) : null;
-        const recent = combatant?.lastBlockMode ?? null;
-        let defaultIdx =
-            recent ?
-                entries.findIndex(
-                    (e) => e.itemId === recent.itemId && e.smId === recent.smId,
-                )
-            :   -1;
-        if (defaultIdx < 0) {
-            defaultIdx = Math.max(
-                0,
-                indexOfBestMastery(entries, (e) => e.ml.constrainedEffective),
-            );
-        }
-
-        // Pick the blocking strike mode + Additional Modifier (dialog, or from
-        // scope when skipDialog: `itemId` + `strikeModeId` + `situationalModifier`).
-        const input = await resolveActionInput<{
-            key: string;
-            situationalModifier: number;
-        }>(context, {
-            fromScope: (s) => {
-                const idx = entries.findIndex(
-                    (e) => e.itemId === s.itemId && e.smId === s.strikeModeId,
-                );
-                return {
-                    key: idx >= 0 ? String(idx) : String(defaultIdx),
-                    situationalModifier:
-                        Number.parseInt(String(s.situationalModifier), 10) || 0,
-                };
-            },
-            dialog: () =>
-                showDefenseDialog(
-                    `${this.name} — Select Block`,
-                    "Block with:",
-                    choices,
-                    String(defaultIdx),
-                ),
-        });
-        if (!input) return;
-        const entry = entries[Number(input.key)];
-        if (!entry) return;
-
-        const defendResult = new DefendResult(
-            {
-                testType: TEST_TYPE.BLOCK.id,
-                // Clone so the defense's situational delta doesn't mutate the
-                // strike mode's live block modifier.
-                masteryLevelModifier: entry.ml.clone<MasteryLevelModifier>(
-                    {},
-                    { parent: this },
-                ),
-                situationalModifier: input.situationalModifier,
-                speaker: context.speaker,
-                token: context.token ?? undefined,
-            } as any,
-            { parent: this },
-        );
-        const combatResult = this.buildCombatResult(
-            attackResult,
-            defendResult,
-            context,
-        );
-        // evaluate() rolls the block on the defender's client, then resolves.
-        await combatResult.evaluate();
-        // Remember this block mode so it defaults next time on this combatant.
-        await combatant?.recordBlockMode(entry.itemId, entry.smId);
-        await this.postCombatResultCard(
-            combatResult,
-            attackResult,
-            `Block w/ ${entry.itemName}`,
-            context,
-        );
-    }
-
-    /**
-     * Resumes automated combat using the Dodge skill. Rolls dodge test and completes processing
-     * of the automated resume (including chat card output). If dodge fails, presents button on
-     * chat card allowing the defender to calculate impact.
-     *
-     * @remarks
-     * One of `combatResult` or `attackResult` must be supplied in `context.scope`:
-     * - `combatResult` is the prior automated resume result that is being reassessed
-     * - `attackResult` is the result of the automated attack that initiated the automated resume
-     *
-     * @param context - Action context carrying the combat or attack result in its scope.
-     * @param [context.scope.priorTestResult] A prior opposed test result that is being retried.
-     * @param [context.scope.attackResult] The test result that initiated the opposed test
-     */
-    async automatedDodgeResume(
-        context: SohlActionContext<Partial<CombatResult.ContextScope>>,
-    ): Promise<void> {
-        const attackResult = this.rehydrateAttackResult(context);
-        if (!attackResult) return;
-
-        // Dodge rolls the defender's Dodge skill; no dialog.
-        const dodgeML = resolveSkillMasteryLevel(this, SKILL_CODE.DODGE);
-        if (!dodgeML) {
-            sohl.log.uiWarn(
-                `${this.name} has no Dodge skill to defend with.`,
-            );
-            return;
-        }
-        const defendResult = new DefendResult(
-            {
-                testType: TEST_TYPE.DODGE.id,
-                // Clone so the defense's situational delta doesn't mutate the
-                // live skill modifier.
-                masteryLevelModifier: dodgeML.clone<MasteryLevelModifier>(
-                    {},
-                    { parent: this },
-                ),
-                situationalModifier: 0,
-                speaker: context.speaker,
-                token: context.token ?? undefined,
-            } as any,
-            { parent: this },
-        );
-        const combatResult = this.buildCombatResult(
-            attackResult,
-            defendResult,
-            context,
-        );
-        // evaluate() rolls the dodge on the defender's client, then resolves the
-        // opposed outcome (the attacker side is read as a snapshot).
-        await combatResult.evaluate();
-        await this.postCombatResultCard(
-            combatResult,
-            attackResult,
-            "Dodge",
-            context,
-        );
-    }
-
-    /**
-     * Resume automated combat with the **Counterstrike** defense — "offense is
-     * the best defense". Unlike Block/Dodge/Ignore, the defender does not roll a
-     * {@link DefendResult}: the counterstrike is itself a second attack (an
-     * {@link AttackResult}) aimed back at the original attacker, so **both** sides
-     * can land in the same exchange. The resolved {@link CombatResult} carries one
-     * "Calculate Injury" button per landing side (the attacker's blow against this
-     * defender, and the counterstrike against the attacker).
-     *
-     * The defender picks a **melee** attack strike mode in reach of the attacker
-     * (the `noAttack` trait excludes a mode; missile modes are never offered — a
-     * counterstrike is a melee reaction, and an out-of-reach attacker yields none)
-     * and a body part to aim at. The default is the most-recently-used attack mode
-     * (a counterstrike *is* an attack), else the best-chance attack mode. Both
-     * choices are bypassable via `scope` (`itemId` + `strikeModeId` +
-     * `situationalModifier`, and `aim`) when `skipDialog` is set.
-     *
-     * The attack snapshot arrives in `context.scope.attackResultJson` (set by the
-     * defender's chat-card button); its speaker token identifies the attacker.
-     *
-     * @param context - Action context carrying the attacker's attack snapshot in its scope.
-     */
-    async automatedCounterstrikeResume(
-        context: SohlActionContext<Partial<CombatResult.ContextScope>>,
-    ): Promise<void> {
-        const attackResult = this.rehydrateAttackResult(context);
-        if (!attackResult) return;
-
-        // This defender's combatant — for the recent-mode default + persistence.
-        const combatant = context.token ? findCombatant(context.token) : null;
-
-        // The counterstrike targets the original attacker (recovered from the
-        // attack snapshot's speaker token); distance is defender → attacker.
-        const rc = resolveCounterstrikeContext(attackResult, combatant);
-        if (!rc) return;
-
-        // Counterstrike uses melee modes within reach of the attacker. A mode is
-        // eligible when it can attack (gated by `noAttack`, via the collector) and
-        // its counterstrike modifier is not independently disabled. A ranged
-        // attacker out of melee reach yields none → cannot counterstrike.
-        const entries = collectAttackableStrikeModes(
-            this,
-            rc.distanceFeet,
-        ).filter(
-            (e) =>
-                (e.strikeMode as any).isMelee &&
-                !(e.strikeMode as any).defense?.counterstrike?.disabled,
-        );
-        if (!entries.length) {
-            sohl.log.uiWarn(
-                `${this.name} has no melee strike mode in reach to counterstrike.`,
-            );
-            return;
-        }
-        const choices: Record<string, string> = {};
-        entries.forEach((e, i) => {
-            choices[String(i)] = `${e.itemName} — ${e.strikeMode.name}`;
-        });
-
-        // Default: the most-recently-used attack mode (a counterstrike shares the
-        // attack-mode history), else the best chance — ranked by the
-        // counterstrike modifier, which is what the counterstrike actually rolls.
-        const recent = combatant?.lastAttackMode ?? null;
-        let defaultIdx =
-            recent ?
-                entries.findIndex(
-                    (e) => e.itemId === recent.itemId && e.smId === recent.smId,
-                )
-            :   -1;
-        if (defaultIdx < 0) {
-            defaultIdx = Math.max(
-                0,
-                indexOfBestMastery(
-                    entries,
-                    (e) =>
-                        (e.strikeMode as any).defense.counterstrike
-                            .constrainedEffective,
-                ),
-            );
-        }
-
-        // Pick the counterstrike strike mode + Additional Modifier.
-        const modeInput = await resolveActionInput<{
-            key: string;
-            situationalModifier: number;
-        }>(context, {
-            fromScope: (s) => {
-                const idx = entries.findIndex(
-                    (e) => e.itemId === s.itemId && e.smId === s.strikeModeId,
-                );
-                return {
-                    key: idx >= 0 ? String(idx) : String(defaultIdx),
-                    situationalModifier:
-                        Number.parseInt(String(s.situationalModifier), 10) || 0,
-                };
-            },
-            dialog: () =>
-                showDefenseDialog(
-                    `${this.name} — Select Counterstrike`,
-                    "Counterstrike with:",
-                    choices,
-                    String(defaultIdx),
-                ),
-        });
-        if (!modeInput) return;
-        const entry = entries[Number(modeInput.key)];
-        if (!entry) return;
-        const sm = entry.strikeMode as any;
-
-        // Aim the counterstrike at a body part on the attacker.
-        const aimChoices = buildAimChoices(rc.attacker.actor);
-        const defaultAim = Object.keys(aimChoices)[0] ?? "";
-        const aim = await resolveActionInput<string | null>(context, {
-            fromScope: (s) => String((s as any).aim ?? defaultAim),
-            dialog: () =>
-                pickChoice(
-                    `${this.name} — Aim Counterstrike`,
-                    "Aim at:",
-                    aimChoices,
-                    defaultAim,
-                ),
-        });
-        if (aim === null) return;
-
-        // The counterstrike is a melee attack by this defender against the
-        // attacker — modelled as an AttackResult in the CombatResult's defender
-        // slot (so both sides can land and carry impact). It rolls the strike
-        // mode's **counterstrike** modifier (same skill base as attack, but its
-        // own deltas) rather than the plain attack modifier.
-        const counter = buildAttackResult({
-            attackML: sm.defense.counterstrike,
-            impact: sm.impact,
-            parent: this,
-            token: context.token ?? null,
-            testType: TEST_TYPE.AUTOCOMBATMELEE.id,
-            aimBodyPartCode: aim,
-            spread: sm.spread?.effective ?? 0,
-            title: entry.itemName,
-        });
-        if (modeInput.situationalModifier) {
-            counter.masteryLevelModifier.add(
-                VALUE_DELTA_INFO.PLAYER,
-                modeInput.situationalModifier,
-            );
-        }
-
-        const combatResult = this.buildCombatResult(
-            attackResult,
-            counter,
-            context,
-        );
-        // evaluate() rolls the counterstrike on the defender's client, then
-        // resolves the exchange and rolls each landing side's impact.
-        await combatResult.evaluate();
-        // A counterstrike is an attack — remember it as the last attack mode.
-        await combatant?.recordAttackMode(entry.itemId, entry.smId);
-
-        if (context.noChat) return;
-        const cardData = buildCombatCardData({
-            combatResult,
-            title: "Attack Result",
-            actorId: this.id,
-            attackerName: attackResult.speaker?.name ?? "",
-            defenderName: this.name,
-            attackWeapon: attackResult.title ?? "",
-            defenseLabel: `Counterstrike w/ ${entry.itemName}`,
-            // The attacker's blow strikes this defender.
-            attackTarget: { name: this.name, actorUuid: this.data.uuid },
-            // The counterstrike strikes the original attacker.
-            defendTarget: rc.attackerAddress,
-        });
-        await context.speaker.toChat(
-            toFilePath("systems/sohl/templates/chat/attack-result-card.hbs"),
-            cardData,
-        );
-    }
-
-    /**
-     * Perform the Ignore defense. Completes processing of the automated resume (including
-     * chat card output). If the defense results in a hit, presents button on chat card
-     * allowing the defender to calculate impact.
-     *
-     * One of `combatResult` or `attackResult` must be supplied in `context.scope`:
-     * - `combatResult` is the prior automated resume result that is being reassessed
-     * - `attackResult` is the result of the automated attack that initiated the automated resume
-     *
-     * @param context - Action context carrying the combat or attack result in its scope.
-     * @param [context.scope.priorTestResult] A prior opposed test result that is being retried.
-     * @param [context.scope.attackResult] The test result that initiated the opposed test
-     */
-    async automatedIgnoreResume(
-        context: SohlActionContext<Partial<CombatResult.ContextScope>>,
-    ): Promise<void> {
-        const attackResult = this.rehydrateAttackResult(context);
-        if (!attackResult) return;
-
-        // Ignore = no defensive contest: a non-rolling placeholder. Resolve
-        // directly (no defender roll) via opposedTestEvaluate.
-        const defendResult = new DefendResult(
-            {
-                testType: TEST_TYPE.IGNORE.id,
-                situationalModifier: 0,
-                speaker: context.speaker,
-                token: context.token ?? undefined,
-            } as any,
-            { parent: this },
-        );
-        const combatResult = this.buildCombatResult(
-            attackResult,
-            defendResult,
-            context,
-        );
-        combatResult.opposedTestEvaluate();
-        await this.postCombatResultCard(
-            combatResult,
-            attackResult,
-            "Ignore",
-            context,
-        );
-    }
-
-    /**
-     * Rehydrate the attacker's evaluated `AttackResult` snapshot from the defense
-     * button's dataset (`data-attack-result-json` → `scope.attackResultJson`).
-     * Returns `null` (with a warning) when absent. Parent is this defender's
-     * logic, per the snapshot-on-defender model.
-     *
-     * @param context - Action context whose scope holds the attack-result JSON snapshot.
-     * @returns The rehydrated `AttackResult`, or null if no snapshot is present.
-     */
-    private rehydrateAttackResult(
-        context: SohlActionContext<Partial<CombatResult.ContextScope>>,
-    ): AttackResult | null {
-        const json = (context.scope as any)?.attackResultJson;
-        if (!json) {
-            sohl.log.uiWarn(
-                `${this.name}: automated-combat resume had no attack result to resolve.`,
-            );
-            return null;
-        }
-        return instanceFromJSON<AttackResult>(json, this);
-    }
-
-    /**
-     * Compose the `CombatResult` for a resolved exchange (attacker snapshot + defender response).
-     * @param attackResult - The attacker's evaluated attack snapshot.
-     * @param defendResult - The defender's response (a defend result or counterstrike attack).
-     * @param context - Action context supplying the speaker for the result.
-     * @returns The composed `CombatResult` pairing the attack and defense.
-     */
-    private buildCombatResult(
-        attackResult: AttackResult,
-        defendResult: AttackResult | DefendResult,
-        context: SohlActionContext<Partial<CombatResult.ContextScope>>,
-    ): CombatResult {
-        return new CombatResult(
-            {
-                attackResult,
-                defendResult,
-                sourceTestResult: attackResult,
-                targetTestResult: defendResult,
-                speaker: context.speaker,
-            } as any,
-            { parent: this },
-        );
-    }
-
-    /**
-     * Post the combat-result card as the defender. A landing blow carries a
-     * "Calculate <Token> Injury" button. Suppressed when `context.noChat`.
-     * @param combatResult - The resolved exchange to render on the card.
-     * @param attackResult - The attacker's attack snapshot for the card data.
-     * @param defenseLabel - Human-readable label describing the defense used.
-     * @param context - Action context supplying the speaker and chat-suppression flag.
-     */
-    private async postCombatResultCard(
-        combatResult: CombatResult,
-        attackResult: AttackResult,
-        defenseLabel: string,
-        context: SohlActionContext<Partial<CombatResult.ContextScope>>,
-    ): Promise<void> {
-        if (context.noChat) return;
-        const cardData = buildCombatCardData({
-            combatResult,
-            title: "Attack Result",
-            actorId: this.id,
-            attackerName: attackResult.speaker?.name ?? "",
-            defenderName: this.name,
-            attackWeapon: attackResult.title ?? "",
-            defenseLabel,
-            attackTarget: { name: this.name, actorUuid: this.data.uuid },
-        });
-        await context.speaker.toChat(
-            toFilePath("systems/sohl/templates/chat/attack-result-card.hbs"),
-            cardData,
-        );
-    }
-
-    /**
      * Define and return all intrinsic actions for this logic type.
      * @returns A map of action shortcodes to their definitions
      */
@@ -1140,46 +646,6 @@ export class BeingLogic<
                 scope: SOHL_ACTION_SCOPE.SELF,
                 iconFAClass: "sohl-continue",
                 executor: "opposedTestResume",
-                visible: "false",
-                group: SOHL_CONTEXT_MENU_SORT_GROUP.HIDDEN,
-            },
-            {
-                shortcode: "automatedBlockResume",
-                subType: ACTION_SUBTYPE.INTRINSIC,
-                title: "SOHL.Being.Action.automatedBlockResume",
-                scope: SOHL_ACTION_SCOPE.SELF,
-                iconFAClass: "sohl-shield-reflect",
-                executor: "automatedBlockResume",
-                visible: "false",
-                group: SOHL_CONTEXT_MENU_SORT_GROUP.HIDDEN,
-            },
-            {
-                shortcode: "automatedCounterstrikeResume",
-                subType: ACTION_SUBTYPE.INTRINSIC,
-                title: "SOHL.Being.Action.automatedCounterstrikeResume",
-                scope: SOHL_ACTION_SCOPE.SELF,
-                iconFAClass: "sohl-riposte",
-                executor: "automatedCounterstrikeResume",
-                visible: "false",
-                group: SOHL_CONTEXT_MENU_SORT_GROUP.HIDDEN,
-            },
-            {
-                shortcode: "automatedDodgeResume",
-                subType: ACTION_SUBTYPE.INTRINSIC,
-                title: "SOHL.Being.Action.automatedDodgeResume",
-                scope: SOHL_ACTION_SCOPE.SELF,
-                iconFAClass: "sohl-dodge",
-                executor: "automatedDodgeResume",
-                visible: "false",
-                group: SOHL_CONTEXT_MENU_SORT_GROUP.HIDDEN,
-            },
-            {
-                shortcode: "automatedIgnoreResume",
-                subType: ACTION_SUBTYPE.INTRINSIC,
-                title: "SOHL.Being.Action.automatedIgnoreResume",
-                scope: SOHL_ACTION_SCOPE.SELF,
-                iconFAClass: "sohl-shrug",
-                executor: "automatedIgnoreResume",
                 visible: "false",
                 group: SOHL_CONTEXT_MENU_SORT_GROUP.HIDDEN,
             },
